Move metrics into reports namespace
