+++ app/views/admin/reports/metrics/index.html.erb
@@ -17,7 +17,7 @@
       <tr class="t-snapshot-table-row">
         <td class="table__cell--middle"><%= snapshot.created_at.strftime('%e %B %Y') %></td>
         <td>
-          <%= link_to 'View', admin_metric_path(snapshot), class: 'btn btn-primary btn-block t-snapshot-view-button' %>
+          <%= link_to 'View', admin_reports_metric_path(snapshot), class: 'btn btn-primary btn-block t-snapshot-view-button' %>
         </td>
       </tr>
     <% end %>
