+++ app/views/layouts/admin.html.erb
@@ -26,7 +26,13 @@
         <%= active_link_to 'Firms', admin_firms_path, wrap_tag: 'li' %>
         <%= active_link_to 'Advisers', admin_advisers_path, wrap_tag: 'li' %>
         <%= active_link_to 'Principals', admin_principals_path, wrap_tag: 'li' %>
-        <%= active_link_to 'Metrics', admin_metrics_path, wrap_tag: 'li' %>
+
+        <li class="<%= 'active' if is_active_link?('/admin/reports', :inclusive) %>">
+          <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-expanded="false">Reports<span class="caret"></span></a>
+          <ul class="dropdown-menu" role="menu">
+            <%= active_link_to 'Metrics', admin_reports_metrics_path, wrap_tag: 'li' %>
+          </ul>
+        </li>
 
         <li class="<%= 'active' if is_active_link?('/admin/lookup', :inclusive) %>">
           <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-expanded="false">Lookup Data <span class="caret"></span></a>
