+++ spec/support/admin/metrics_show_page.rb
@@ -1,7 +1,7 @@
 module Admin
   class MetricsShowPage < SitePrism::Page
-    set_url '/admin/metrics/{snapshot_id}'
-    set_url_matcher %r{/admin/metrics/[0-9]+}
+    set_url '/admin/reports/metrics/{snapshot_id}'
+    set_url_matcher %r{/admin/reports/metrics/[0-9]+}
 
     element :table, '.t-table'
   end
